List of changes from the community along with my own improvements.
Allow --name to be modified

Name wasn't modifiable; this fixes it to work like the `--description`
flag, i.e.

    sogs --rooms some-room --name 'Some New Name'

now works.

Add `/rooms/all/<SessionID:sid>` for server-wide message deletion.

This causes all messages by the given Session id to be deleted from the
server. The deleting user must be a global moderator on the server.

This facility is needed to allow a server-wide 'Ban and Delete' feature
for Session moderators and admins.

Change sogs.ini.sample to 10MB

Change to 10MB in config.py

Just correct the nginx config proxy_pass pattern in comments

Notes:

When use the ` http://127.3.2.1:4242$request_uri ` in the nginx config
file `sogs-proxy`, it will redirect to error page `
http://127.3.2.1:4242/r/room_token ` by clicking the room link on the
SOGS index web page.

When use the ` http://127.3.2.1:4242/ ` instead, it will redirect to `
http://example.com/r/room_token/ ` or ` http://<public_ip>/r/room_token
`, and this works well for the SOGS web accessing.

I also commit this fix and make a PR to the deb package building branch
` debian/sid `.

Added setting to block private messages from users to users. Excluding admins/moderators.

Minor change in config.py

Cleared up unnecessary things and inherited the latest changes in the repository.